Created framework for item creator

+++ Item/TemperatureConverter.cs
@@ -9,9 +9,9 @@ namespace ItemEvaluator
 	static class TemperatureConverter
 	{
 		//Values are saved in database as fahrenheit for all temperatures. Conversion needed to save and retrieve
-		public static float SaveValueTemperature(TemperatureScale temperatureScale, float temperature)
+		public static double SaveValueTemperature(TemperatureScale temperatureScale, float temperature)
 		{
-			float saveTemp;
+			double saveTemp;
 			if (temperatureScale == TemperatureScale.Fahrenheit)
 				saveTemp = temperature;
 			else if (temperatureScale == TemperatureScale.Celsius)
@@ -23,7 +23,7 @@ namespace ItemEvaluator
 				
 		public static string DisplayValueTemperature(TemperatureScale temperatureScale, float temperature)
 		{
-			float tempValue;
+			double tempValue;
 			string displayValue;
 			if (temperatureScale == TemperatureScale.Fahrenheit)
 			{
@@ -43,55 +43,55 @@ namespace ItemEvaluator
 			return displayValue;
 		}
 
-		private static float FahrenheitToCelsius(float fahrenheitTemp)
+		private static double FahrenheitToCelsius(double fahrenheitTemp)
 		{
-			float celciusTemp = (5 / 9) * (fahrenheitTemp - 32f);
+			double celciusTemp = (5 / 9) * (fahrenheitTemp - 32f);
 			return celciusTemp;
 		}
 
-		private static float FahrenheitToKelvin(float fahrenheitTemp)
+		private static double FahrenheitToKelvin(double fahrenheitTemp)
 		{
-			float kelvinTemp = ((5 / 9) * (fahrenheitTemp - 32f)) + 273f;
+			double kelvinTemp = ((5 / 9) * (fahrenheitTemp - 32f)) + 273f;
 			return kelvinTemp;
 		}
 
-		private static float CelsiusToFahrenheit(float celsiusTemp)
+		private static double CelsiusToFahrenheit(double celsiusTemp)
 		{
-			float fahrenheitTemp = ((9 / 5) * celsiusTemp) + 32f;
+			double fahrenheitTemp = ((9 / 5) * celsiusTemp) + 32f;
 			return fahrenheitTemp;
 		}
 
-		private static float CelsiusToKelvin(float celsiusTemp)
+		private static double CelsiusToKelvin(double celsiusTemp)
 		{
-			float kelvinTemp = celsiusTemp + 273f;
+			double kelvinTemp = celsiusTemp + 273f;
 			return kelvinTemp;
 		}
 
-		private static float KelvinToFahrenheit(float kelvinTemp)
+		private static double KelvinToFahrenheit(double kelvinTemp)
 		{
-			float fahrenheitTemp = ((9 / 5) * (kelvinTemp - 273f)) + 32f;
+			double fahrenheitTemp = ((9 / 5) * (kelvinTemp - 273f)) + 32f;
 			return fahrenheitTemp;
 		}
 
-		private static float KelvinToCelsius(float kelvinTemp)
+		private static double KelvinToCelsius(double kelvinTemp)
 		{
-			float celsiusTemp = kelvinTemp - 273f;
+			double celsiusTemp = kelvinTemp - 273f;
 			return celsiusTemp;
 		}
 
-		private static string DisplayFahrenheit(float fahrenheitTemp)
+		private static string DisplayFahrenheit(double fahrenheitTemp)
 		{
 			string displayString = $"{fahrenheitTemp}°F";
 			return displayString;
 		}
 
-		private static string DisplayCelsius(float celsiusTemp)
+		private static string DisplayCelsius(double celsiusTemp)
 		{
 			string displayString = $"{celsiusTemp}°C";
 			return displayString;
 		}
 
-		private static string DisplayKelvin(float kelvinTemp)
+		private static string DisplayKelvin(double kelvinTemp)
 		{
 			string displayString = $"{kelvinTemp}K";
 			return displayString;
